% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{CapWords}
\alias{CapWords}
\title{Capitalize Words}
\usage{
CapWords(s, strict = FALSE)
}
\arguments{
\item{s}{string}

\item{strict}{logical, indicating whether or not string it set to title case , Default: FALSE}
}
\value{
returns capitalized string
}
\description{
capitalize the first letter in each words in a string
}
\examples{
 CapWords("example eXAMPLE", FALSE)
 # [1] "Example EXAMPLE"
 CapWords("example eXAMPLE", TRUE)
 # [1] "Example Example"
}
