% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cats.R
\docType{data}
\name{Cats}
\alias{Cats}
\title{Dataset with Cats}
\format{
A data frame with 2000 rows and 4 variables:
\describe{
  \item{\code{Reward}}{integer Food or Affection}
  \item{\code{Dance}}{integer Yes or No}
  \item{\code{Alignment}}{integer Good or Evil}
  \item{\code{Ratings}}{double Cats rate their owners (average of multiple seven-point Likert-type scale (1 = Hate ... 7 = Love)} 
}
}
\usage{
Cats
}
\description{
Shamelessly adapted from Field (2017).
}
\details{
Example data for BFW
}
\keyword{datasets}
