% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{ChangeNames}
\alias{ChangeNames}
\title{Change Names}
\usage{
ChangeNames(
  x,
  names,
  single.items = FALSE,
  row.names = FALSE,
  param = NULL,
  where = NULL,
  environment = NULL
)
}
\arguments{
\item{x}{list, vector, matrix, dataframe or a list of such items}

\item{names}{names to insert}

\item{single.items}{logical, indicating whether or not to use names rather than colnames or rownames, Default: FALSE}

\item{row.names}{logical, indicating whether or not to use rownames rather than colnames, Default: FALSE}

\item{param}{Variable name, Default: NULL}

\item{where}{select parents, Default: NULL}

\item{environment}{select reference environment, Default: NULL}
}
\value{
returns Named items
 # ABC <- c("1","2","3")
 # "1" "2" "3"
 # ChangeNames(ABC, names = c("A","B","C") , single.items = TRUE)
 #  A   B   C 
 # "1" "2" "3"
}
\description{
Change names, colnames or rownames of single items or a list of items
}
