% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_misc.R
\name{ContrastNames}
\alias{ContrastNames}
\title{Contrast Names}
\usage{
ContrastNames(items, job.names, col.names)
}
\arguments{
\item{items}{items to create names for}

\item{job.names}{names of all parameters in analysis, Default: NULL}

\item{col.names}{columns in MCMC to create names from}
}
\description{
utilize the AddNames function to create contrast names
}
