% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{DistinctColors}
\alias{DistinctColors}
\title{Distinct Colors}
\usage{
DistinctColors(range, random = FALSE)
}
\arguments{
\item{range}{number of colors as sequence}

\item{random}{logical, indicating whether or not to provide random colors, Default: FALSE}
}
\description{
create vector containing Hex color codes
}
\examples{
 DistinctColors(1:3)
 # [1] "#FFFF00" "#1CE6FF" "#FF34FF"
 set.seed(1)
 DistinctColors(1:3, TRUE)
 # [1] "#575329" "#CB7E98" "#D86A78"
}
