% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{ETA}
\alias{ETA}
\title{ETA}
\usage{
ETA(start.time, i, total)
}
\arguments{
\item{start.time}{Start time (preset variable with Sys.time())}

\item{i}{incremental steps towards total}

\item{total}{Total number of steps}
}
\description{
Print estimated time for arrival (ETA)
}
\seealso{
\code{\link[utils]{flush.console}}
}
