% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{FileName}
\alias{FileName}
\title{File Name}
\usage{
FileName(
  project = "Project",
  subset = NULL,
  type = NULL,
  name = NULL,
  unix = TRUE,
  ...
)
}
\arguments{
\item{project}{name of project, Default: 'Project'}

\item{subset}{define subset of data, Default: NULL}

\item{type}{type of data, Default: NULL}

\item{name}{save name, Default: NULL}

\item{unix}{logical, indicating whether or not to append unix timestamp, Default: TRUE}

\item{...}{further arguments passed to or from other methods}
}
\description{
simple function to construct a file name for data
}
\examples{
 FileName()
 # [1] "Project-Name-1528834963"

 FileName(project = "Project" ,
         subset = "subset" ,
         type = "longitudinal" ,
         name = "cheese",
         unix = FALSE)
 # [1] "Projectsubset-longitudinal-cheese"
}
