% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{FindEnvironment}
\alias{FindEnvironment}
\title{Find Environment}
\usage{
FindEnvironment(x, where = NULL)
}
\arguments{
\item{x}{any type of named object}

\item{where}{select reference environment, Default: NULL}
}
\value{
returns Found environment, Default: R_GlobalEnv.
}
\description{
Find the environment of a selected variable.
}
