% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{FlattenList}
\alias{FlattenList}
\title{Flatten List}
\usage{
FlattenList(li, rm.duplicated = TRUE, unname.li = TRUE, rm.empty = TRUE)
}
\arguments{
\item{li}{list to flatten}

\item{rm.duplicated}{logical, indicating whether or not to remove duplicated lists, Default: TRUE}

\item{unname.li}{logical, indicating whether or not to unname lists, Default: TRUE}

\item{rm.empty}{logical, indicating whether or not to remove empty lists, Default: TRUE}
}
\description{
flatten a nested list into a single list
}
