% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{GammaDist}
\alias{GammaDist}
\title{Gamma Distribution}
\usage{
GammaDist(mode, sd)
}
\arguments{
\item{mode}{mode from data}

\item{sd}{standard deviation from data}
}
\description{
compute gamma distribution (shape and rate) from mode and standard deviation
}
\examples{
 GammaDist(1,0.5)
 # $shape
 # [1] 5.828427
 # $rate
 # [1] 4.828427
}
