% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Interleave}
\alias{Interleave}
\title{Interleave}
\usage{
Interleave(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
mixed vector
}
\description{
mix vectors by alternating between them
}
\examples{
 a <- 1:3
 b <- LETTERS[1:3]
 Interleave(a,b)
 # [1] "1" "A" "2" "B" "3" "C"
}
