% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Layout}
\alias{Layout}
\title{Layout}
\usage{
Layout(x = "a4", layout.inverse = FALSE)
}
\arguments{
\item{x}{type of layout, Default: 'a4'}

\item{layout.inverse}{logical, indicating whether or not to inverse layout (e.g., landscape) , Default: FALSE}
}
\value{
width and height of select medium
}
\description{
collection of layout sizes
}
\examples{
 Layout()
 # [1]  8.3 11.7
}
