% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{MultiGrep}
\alias{MultiGrep}
\title{Multi Grep}
\usage{
MultiGrep(find, from, remove = NULL, value = TRUE)
}
\arguments{
\item{find}{vector to find}

\item{from}{vector to find from}

\item{remove}{variables to remove, Default: NULL}

\item{value}{logical, if TRUE returns value, Default: TRUE}
}
\description{
Use multiple patterns from vector to find element in another vector, with option to remove certain patterns
}
