% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Normalize}
\alias{Normalize}
\title{Normalize}
\usage{
Normalize(x)
}
\arguments{
\item{x}{numeric vector to normalize}
}
\description{
simple function to normalize data
}
\examples{
\dontrun{
if(interactive()){
 Normalize(1:10)
 # [1] 0.01818182 0.03636364 0.05454545 0.07272727 0.09090909 
       0.10909091 0.12727273 0.14545455 0.16363636 0.18181818
 }
}
}
