% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{PadVector}
\alias{PadVector}
\title{Pad Vector}
\usage{
PadVector(v)
}
\arguments{
\item{v}{numeric vector to pad}
}
\description{
Pad a numeric vector according to the highest value
}
\examples{
 PadVector(1:10)
 # [1] "01" "02" "03" "04" "05" "06" "07" "08" "09" "10"
}
