% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{PaperSize}
\alias{PaperSize}
\title{Paper Size}
\usage{
PaperSize(x = "a4")
}
\arguments{
\item{x}{select medium to print on, Default: 'a4'}
}
\value{
width and height of select medium
}
\description{
collection of print sizes
}
\examples{
 PaperSize()
 # [1]  8.3 11.7
}
