% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_circlize.R
\name{PlotCirclize}
\alias{PlotCirclize}
\title{Circlize Plot}
\usage{
PlotCirclize(
  data,
  category.spacing = 1.2,
  category.inset = c(-0.4, 0),
  monochrome = TRUE,
  plot.colors = c("#CCCCCC", "#DEDEDE"),
  font.type = "serif"
)
}
\arguments{
\item{data}{data for circlize plot}

\item{category.spacing}{spacing between category items , Default: 1.25}

\item{category.inset}{inset of category items form plot , Default: c(-0.5, 0)}

\item{monochrome}{logical, indicating whether or not to use monochrome colors, else use \link[bfw]{DistinctColors}, Default: TRUE}

\item{plot.colors}{range of color to use, Default: c("#CCCCCC", "#DEDEDE")}

\item{font.type}{font type used for visualizations, Default: 'serif'}
}
\description{
Create a circlize plot
}
\seealso{
\code{\link[grDevices]{dev}},
\code{\link[grDevices]{recordPlot}}
\code{\link[graphics]{legend}}
\code{\link[circlize]{circos.par}},
\code{\link[circlize]{chordDiagram}},
\code{\link[circlize]{circos.trackPlotRegion}},
\code{\link[circlize]{circos.clear}}
}
