% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mean.R
\name{PlotMean}
\alias{PlotMean}
\title{Plot Mean}
\usage{
PlotMean(
  data,
  monochrome = TRUE,
  plot.colors = c("#495054", "#e3e8ea"),
  font.type = "serif",
  run.repeated = FALSE,
  run.split = FALSE,
  y.split = FALSE,
  ribbon.plot = TRUE,
  y.text = "Score",
  x.text = NULL,
  remove.x = FALSE
)
}
\arguments{
\item{data}{MCMC data to plot}

\item{monochrome}{logical, indicating whether or not to use monochrome colors, else use \link[bfw]{DistinctColors}, Default: TRUE}

\item{plot.colors}{range of color to use, Default: c("#495054", "#e3e8ea")}

\item{font.type}{font type used for visualizations, Default: 'serif'}

\item{run.repeated}{logical, indicating whether or not to use repeated measures plot, Default: FALSE}

\item{run.split}{logical, indicating whether or not to use split violin plot and compare distribution between groups, Default: FALSE}

\item{y.split}{logical, indicating whether or not to split within (TRUE) or between groups, Default: FALSE}

\item{ribbon.plot}{logical, indicating whether or not to use ribbon plot for HDI, Default: TRUE}

\item{y.text}{label on y axis, Default: 'Score'}

\item{x.text}{label on x axis, Default: NULL}

\item{remove.x}{logical, indicating whether or not to show x.axis information, Default: FALSE}
}
\description{
Create a (repeated) mean plot
}
\seealso{
\code{\link[ggplot2]{ggproto}},
 \code{\link[ggplot2]{ggplot2-ggproto}},
 \code{\link[ggplot2]{aes}},
 \code{\link[ggplot2]{margin}},
 \code{\link[ggplot2]{geom_boxplot}},
 \code{\link[ggplot2]{geom_crossbar}},
 \code{\link[ggplot2]{geom_path}},
 \code{\link[ggplot2]{geom_ribbon}},
 \code{\link[ggplot2]{geom_violin}},
 \code{\link[ggplot2]{ggplot}},
 \code{\link[ggplot2]{scale_manual}},
 \code{\link[ggplot2]{scale_x_discrete}},
 \code{\link[ggplot2]{theme}},
 \code{\link[ggplot2]{layer}},
 \code{\link[ggplot2]{labs}}
 \code{\link[plyr]{arrange}},
 \code{\link[plyr]{rbind.fill}}
 \code{\link[scales]{zero_range}}
 \code{\link[grid]{grid.grob}},
 \code{\link[grid]{grobName}},
 \code{\link[grid]{unit}}
 \code{\link[stats]{approxfun}}
 \code{\link[grDevices]{colorRamp}}
}
