% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param.R
\name{PlotParam}
\alias{PlotParam}
\title{Plot Param}
\usage{
PlotParam(data, param, ROPE = FALSE, monochrome = TRUE,
  plot.colors = c("#495054", "#e3e8ea"), font.type = "serif",
  font.size = 4.5, rope.line = -0.2, rope.tick = -0.1,
  rope.label = -0.35, line.size = 0.5, dens.zero.col = "black",
  dens.mean.col = "white", dens.median.col = "white",
  dens.mode.col = "black", dens.rope.col = "black")
}
\arguments{
\item{data}{MCMC data to plot}

\item{param}{parameter of interest}

\item{ROPE}{plot ROPE values, Default: FALSE}

\item{monochrome}{logical, indicating whether or not to use monochrome colors, else use \link[bfw]{DistinctColors}, Default: TRUE}

\item{plot.colors}{range of color to use, Default: c("#495054", "#e3e8ea")}

\item{font.type}{font type used for visualizations, Default: 'serif'}

\item{font.size}{font size, Default: 4.5}

\item{rope.line}{size of ROPE lien, Default: -0.2}

\item{rope.tick}{distance to ROPE tick, Default: -0.1}

\item{rope.label}{distance to ROPE label, Default: -0.35}

\item{line.size}{overall line size, Default: 0.5}

\item{dens.zero.col}{colour of line indicating zero, Default: 'black'}

\item{dens.mean.col}{colour of line indicating mean value, Default: 'white'}

\item{dens.median.col}{colour of line indicating median value, Default: 'white'}

\item{dens.mode.col}{colour of line indicating mode value, Default: 'black'}

\item{dens.rope.col}{colour of line indicating ROPE value, Default: 'black'}
}
\value{
Density plot of parameter values
}
\description{
Create a density plot with parameter values
}
