% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{RemoveEmpty}
\alias{RemoveEmpty}
\title{Remove Empty}
\usage{
RemoveEmpty(x)
}
\arguments{
\item{x}{vector to eliminate NA and blanks}
}
\description{
Remove empty elements in vector
}
\examples{
 RemoveEmpty( c("",NA,"","Remains") )
 # [1] "Remains"
}
