% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{RemoveSpaces}
\alias{RemoveSpaces}
\title{Remove Spaces}
\usage{
RemoveSpaces(x)
}
\arguments{
\item{x}{string}
}
\description{
simple function to remove whitespace
}
\examples{
 RemoveSpaces("  No More S p a c e s")
 # [1] "NoMoreSpaces"
}
