% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_misc.R
\name{RunContrasts}
\alias{RunContrasts}
\title{Run Contrasts}
\usage{
RunContrasts(contrast.type, q.levels, use.contrast, contrasts, data, job.names)
}
\arguments{
\item{contrast.type}{type of contrast: "m" indicate means and standard deviations, "o" indicate frequency}

\item{q.levels}{Number of levels of each variable/column}

\item{use.contrast}{choose from "between", "within" and "mixed". Between compare groups at different conditions. Within compare a group at different conditions. Mixed compute all comparisons}

\item{contrasts}{specified contrasts columns}

\item{data}{data to compute contrasts from}

\item{job.names}{names of all parameters in analysis, Default: NULL}
}
\description{
Compute contrasts from mean and standard deviation (Cohen's d) or frequencies (odds ratio)
}
\seealso{
\code{\link[utils]{combn}}
}
