% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SaveName}
\alias{SaveName}
\title{Save Name}
\usage{
SaveName(project = "Project", dataset = "AllData", type = "Model",
  name = "Name", unix = TRUE, ...)
}
\arguments{
\item{project}{name of project, Default: 'Project'}

\item{dataset}{define subset of data, Default: 'AllData'}

\item{type}{type of data, Default: 'Model'}

\item{name}{save name, Default: 'Name'}

\item{unix}{logical, indicating whether or not to add unix timpestamp to save name, Default: TRUE}

\item{...}{further arguments passed to or from other methods}
}
\description{
simple function to construct a save name for data
}
\examples{
\dontrun{
if(interactive()){
 SaveName()
 # [1] "ProjectAllData-Model-Name-1528408490"
 }
}
}
