% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SavePlot}
\alias{SavePlot}
\title{Save Plot}
\usage{
SavePlot(plot.data, file = "Results/Generic", graphic.type = "pptx",
  plot.size = "15,10", scaling = 100, plot.aspect = NULL,
  save.data = FALSE, vector.graphic = TRUE, point.size = 15,
  font.type = "serif", one.file = TRUE, ppi = 300, units = "in",
  paper = "pw", ...)
}
\arguments{
\item{plot.data}{a list of plots}

\item{file}{save name, Default: 'Results/Generic'}

\item{graphic.type}{type of graphics to use (e.g., pdf, png, ps), Default: 'pptx'}

\item{plot.size}{size of plot, Default: '15,10'}

\item{scaling}{scale size of plot, Default: 100}

\item{plot.aspect}{aspect of plot, Default: NULL}

\item{save.data}{logical, indicating whether or not to save data, Default: FALSE}

\item{vector.graphic}{logical, indicating whether or not visualizations should be vector or raster graphics, Default: TRUE}

\item{point.size}{point size used for visualizations, Default: 15}

\item{font.type}{font type used for visualizations, Default: 'serif'}

\item{one.file}{logical, indicating whether or not visualizations should be placed in one or several files, Default: TRUE}

\item{ppi}{define pixel per inch used for visualizations, Default: 300}

\item{units}{define unit of length used for visualizations, Default: 'in'}

\item{paper}{define a print size for visualizations, Default: 'pw'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Display and/or save plots
}
\seealso{
\code{\link[grDevices]{dev.new}},\code{\link[grDevices]{recordPlot}},\code{\link[grDevices]{setEPS}},\code{\link[grDevices]{graphics.off}},\code{\link[grDevices]{dev.list}},\code{\link[grDevices]{dev.off}}
 \code{\link[ReporteRs]{pptx}},\code{\link[ReporteRs]{addSlide}},\code{\link[ReporteRs]{addPlot}},\code{\link[ReporteRs]{writeDoc}}
}
