% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SingleString}
\alias{SingleString}
\title{Single String}
\usage{
SingleString(x)
}
\arguments{
\item{x}{string}
}
\value{
true or false
}
\description{
determine whether input is a single string
}
\examples{
A <- "This is a single string"
SingleString(A)
# [1] TRUE
is.character(A)
# [1] TRUE
B <- c("This is a vector" , "containing two strings")
SingleString(B)
# [1] FALSE
is.character(B)
# [1] TRUE
}
