% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_fit.R
\name{StatsFit}
\alias{StatsFit}
\title{Fit Data}
\usage{
StatsFit(
  latent = NULL,
  latent.names = NULL,
  observed = NULL,
  observed.names = NULL,
  additional = NULL,
  additional.names = NULL,
  DF,
  params = NULL,
  job.group = NULL,
  initial.list = list(),
  model.name,
  jags.model,
  custom.model = NULL,
  run.ppp = FALSE,
  run.robust = FALSE,
  ...
)
}
\arguments{
\item{latent}{latenr variables, Default: NULL}

\item{latent.names}{optional names for for latent variables, Default: NULL}

\item{observed}{observed variable(s), Default: NULL}

\item{observed.names}{optional names for for observed variable(s), Default: NULL}

\item{additional}{supplemental parameters for fitted data (e.g., indirect pathways and total effect), Default: NULL}

\item{additional.names}{optional names for supplemental parameters, Default: NULL}

\item{DF}{data to analyze}

\item{params}{define parameters to observe, Default: NULL}

\item{job.group}{for some hierarchical models with several layers of parameter names (e.g., latent and observed parameters), Default: NULL}

\item{initial.list}{initial values for analysis, Default: list()}

\item{model.name}{name of model used}

\item{jags.model}{specify which module to use}

\item{custom.model}{define a custom model to use (e.g., string or text file (.txt), Default: NULL}

\item{run.ppp}{logical, indicating whether or not to conduct ppp analysis, Default: FALSE}

\item{run.robust}{logical, indicating whether or not robust analysis, Default: FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Apply latent or observed models to fit data (e.g., SEM, CFA, mediation)
}
\seealso{
\code{\link[stats]{complete.cases}}
}
