% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_regression.R
\name{StatsRegression}
\alias{StatsRegression}
\title{Regression}
\usage{
StatsRegression(
  y = NULL,
  y.names = NULL,
  x = NULL,
  x.names = NULL,
  x.steps = NULL,
  x.blocks = NULL,
  DF,
  params = NULL,
  job.group = NULL,
  initial.list = list(),
  ...
)
}
\arguments{
\item{y}{criterion variable(s), Default: NULL}

\item{y.names}{optional names for criterion variable(s), Default: NULL}

\item{x}{predictor variable(s), Default: NULL}

\item{x.names}{optional names for predictor variable(s), Default: NULL}

\item{x.steps}{define number of steps in hierarchical regression , Default: NULL}

\item{x.blocks}{define which predictors are included in each step (e.g., for three steps "1,2,3") , Default: NULL}

\item{DF}{data to analyze}

\item{params}{define parameters to observe, Default: NULL}

\item{job.group}{for some hierarchical models with several layers of parameter names (e.g., latent and observed parameters), Default: NULL}

\item{initial.list}{initial values for analysis, Default: list()}

\item{...}{further arguments passed to or from other methods}
}
\description{
Simple, multiple and hierarchical regression
}
\seealso{
\code{\link[stats]{complete.cases}}
}
