% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_misc.R
\name{SumToZero}
\alias{SumToZero}
\title{Sum to Zero}
\usage{
SumToZero(q.levels, data, contrasts)
}
\arguments{
\item{q.levels}{number of levels of each variable/column}

\item{data}{data matrix to combine from}

\item{contrasts}{specified contrasts columns}
}
\description{
Compute sum to zero values across all levels of a data matrix
}
\examples{
 data <- matrix(c(1,2),ncol=2)
 colnames(data) <- c("m1[1]","m1[2]")
 SumToZero( 2 , data , contrasts = NULL )
 #               b0[1] b1[1] b1[2]
 #       m1[1]   1.5  -0.5   0.5
}
