% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{TidyCode}
\alias{TidyCode}
\title{Tidy Code}
\usage{
TidyCode(tidy.code, jags = TRUE)
}
\arguments{
\item{tidy.code}{Messy code that needs cleaning}

\item{jags}{logical, if TRUE run code as JAGS model, Default: TRUE}
}
\value{
(Somewhat) tidy code
}
\description{
Small function that clears up messy code
}
\examples{
messy <- "code <- function( x ) {
print (x ) }"
cat(messy)
code <- function( x ) {
print (x ) }
cat ( TidyCode(messy, jags = FALSE) )
code <- function(x) {
   print(x)
}
}
