% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Trim}
\alias{Trim}
\title{Trim}
\usage{
Trim(s, multi = TRUE)
}
\arguments{
\item{s}{string}

\item{multi}{logical, indicating whether or not to remove excess whitespace between characters, Default: TRUE}
}
\description{
remove excess whitespace from string
}
\examples{
 Trim("             Trimmed      string")
 # [1] "Trimmed string"
 Trim("             Trimmed      string", FALSE)
 # [1] "Trimmed      string"
}
