% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{TrimSplit}
\alias{TrimSplit}
\title{Trim Split}
\usage{
TrimSplit(
  x,
  sep = ",",
  fixed = FALSE,
  perl = FALSE,
  useBytes = FALSE,
  rm.empty = TRUE
)
}
\arguments{
\item{x}{string}

\item{sep}{symbol to separate data (e.g., comma-delimited), Default: ','}

\item{fixed}{logical, if TRUE match split exactly, otherwise use regular expressions. Has priority over perl, Default: FALSE}

\item{perl}{logical, indicating whether or not to use Perl-compatible regexps, Default: FALSE}

\item{useBytes}{logical, if TRUE the matching is done byte-by-byte rather than character-by-character, Default: FALSE}

\item{rm.empty}{logical. indicating whether or not to remove empty elements, Default: TRUE}
}
\description{
Extends strsplit by trimming and unlisting string
}
\details{
\link[base]{strsplit}
}
\examples{
 TrimSplit("Data 1,     Data2, Data3")
 # [1] "Data 1" "Data2"  "Data3"
}
