% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{VectorSub}
\alias{VectorSub}
\title{Pattern Matching and Replacement From Vectors}
\usage{
VectorSub(pattern, replacement, string)
}
\arguments{
\item{pattern}{vector containing words to match}

\item{replacement}{vector containing words to replace existing words.}

\item{string}{string to replace from}
}
\value{
modified string with replaced values
}
\description{
extending gsub by matching pattern and replacement from two vectors
}
\examples{
 pattern <- c("A","B","C")
 replacement <- 1:3
 string <- "A went to B went to C"
 VectorSub(pattern,replacement,string)
 # [1] "1 went to 2 went to 3"
}
