% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{ETA}
\alias{ETA}
\title{ETA}
\usage{
ETA(start.time, i, total, results = NULL)
}
\arguments{
\item{start.time}{start time (preset variable with Sys.time())}

\item{i}{incremental steps towards total}

\item{total}{total number of steps}

\item{results}{message to display, Default: NULL}
}
\description{
Print estimated time for arrival (ETA)
}
\seealso{
\code{\link[utils]{flush.console}}
}
