% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_misc.R
\name{MergeMCMC}
\alias{MergeMCMC}
\title{Merge MCMC}
\usage{
MergeMCMC(pat, project.dir = "Results/", data.sets)
}
\arguments{
\item{pat}{pattern to select MCMC chain from}

\item{project.dir}{define where to save data, Default: 'Results/'}

\item{data.sets}{data sets to combine}
}
\value{
Merged MCMC chains
}
\description{
Merge two or more MCMC simulations
}
\seealso{
\code{\link[utils]{head}}
 \code{\link[runjags]{combine.mcmc}}
}
