% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{Normalize}
\alias{Normalize}
\title{Normalize}
\usage{
Normalize(x)
}
\arguments{
\item{x}{numeric vector to normalize}
}
\description{
simple function to normalize data
}
\examples{
Normalize(1:10)
# [1] 0.0182 0.0364 0.0545 0.0727 0.0909
# 0.1091 0.1273 0.1455 0.1636 0.1818
}
