% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{ParseNumber}
\alias{ParseNumber}
\title{Parse Numbers}
\usage{
ParseNumber(x, digits = FALSE)
}
\arguments{
\item{x}{string or vector}

\item{digits}{logical, indicating whether or not to extract decimals, Default: FALSE}
}
\description{
simple function to extract numbers from string/vector
}
\examples{
 ParseNumber("String1WithNumbers2")
 # [1] 1 2
}
\seealso{
\code{\link[stats]{na.omit}}
}
