% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PlotData}
\alias{PlotData}
\title{Plot Data}
\usage{
PlotData(data, data.type = "Mean", ...)
}
\arguments{
\item{data}{data to plot data from}

\item{data.type}{define what kind of data is being used, Default: 'Mean'}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plot data as violin plot visualizing density, box plots to display HDI, whiskers to display standard deviation
}
