% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_mean.R
\name{StatsMean}
\alias{StatsMean}
\title{Mean Data}
\usage{
StatsMean(
  y = NULL,
  y.names = NULL,
  x = NULL,
  x.names = NULL,
  DF,
  params = NULL,
  initial.list = list(),
  ...
)
}
\arguments{
\item{y}{criterion variable(s), Default: NULL}

\item{y.names}{optional names for criterion variable(s), Default: NULL}

\item{x}{categorical variable(s), Default: NULL}

\item{x.names}{optional names for categorical variable(s), Default: NULL}

\item{DF}{User defined data frame, Default: NULL}

\item{params}{define parameters to observe, Default: NULL}

\item{initial.list}{Initial values for simulations, Default: list()}

\item{...}{further arguments passed to or from other methods}
}
\value{
mean and standard deviation
}
\description{
Compute means and standard deviations.
}
