% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_metric.R
\name{StatsMetric}
\alias{StatsMetric}
\title{Predict Metric}
\usage{
StatsMetric(
  y = NULL,
  y.names = NULL,
  x = NULL,
  x.names = NULL,
  DF,
  params = NULL,
  job.group = NULL,
  initial.list = list(),
  model.name,
  jags.model,
  custom.model = NULL,
  run.robust = FALSE,
  ...
)
}
\arguments{
\item{y}{criterion variable(s), Default: NULL}

\item{y.names}{optional names for criterion variable(s), Default: NULL}

\item{x}{categorical variable(s), Default: NULL}

\item{x.names}{optional names for categorical variable(s), Default: NULL}

\item{DF}{data to analyze}

\item{params}{define parameters to observe, Default: NULL}

\item{job.group}{for some hierarchical models with several layers of parameter names (e.g., latent and observed parameters), Default: NULL}

\item{initial.list}{initial values for analysis, Default: list()}

\item{model.name}{name of model used}

\item{jags.model}{specify which module to use}

\item{custom.model}{define a custom model to use (e.g., string or text file (.txt), Default: NULL}

\item{run.robust}{logical, indicating whether or not robust analysis, Default: FALSE}

\item{...}{further arguments passed to or from other methods}
}
\description{
Bayesian alternative to ANOVA
}
\seealso{
\code{\link[stats]{complete.cases}},
\code{\link[stats]{sd}},
\code{\link[stats]{aggregate}},
\code{\link[stats]{median}}
\code{\link[utils]{head}}
}
